pyproject.toml
[project]
name = "fund-advisor"
version = "0.1.0"
description = "个人中国公募基金投资决策助手（单用户自用，仅供信息聚合与辅助分析，不构成投资建议）"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "ShawnHuang" }]
license = { text = "Proprietary (personal use)" }

dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "streamlit>=1.34",
    "plotly>=5.20",
    "pandas>=2.2",
    "loguru>=0.7",
    "python-dateutil>=2.9",
    "akshare>=1.18.56",
    "openai>=2.32.0",
    "tenacity>=9.1.4",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fund_advisor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py311"
